+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 **Merged pull requests:**
 
 - Release 2.4.3 🐎 ([belomaxorka](https://github.com/belomaxorka))
+- Improved app debug [\#1438](https://github.com/torrentpier/torrentpier/pull/1438) ([belomaxorka](https://github.com/belomaxorka))
 - Minor improvements [\#1435](https://github.com/torrentpier/torrentpier/pull/1435) ([belomaxorka](https://github.com/belomaxorka))
 
 ## [v2.4.2](https://github.com/torrentpier/torrentpier/tree/v2.4.2) (2024-03-30)